+++ RELEASENOTES.md
@@ -54,6 +54,9 @@
         operations to the service's lifecycle
         ([#187](https://github.com/androidx/media/issues/187)).
 *   UI:
+    *   Add `ProgressSlider` Composable to `media3-ui-compose-material3` for
+        displaying player progress and performing seeks using dragging and
+        tapping ([#2288](https://github.com/androidx/media/issues/2288)).
 *   Downloads:
     *   Fix potential infinite loops when a `PriorityTooLowException` is handled
         by `SegmentDownloader` (for DASH, HLS and SmoothStreaming). Custom